+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Show clearer error if "rules" is missing from config.
+
 ## 0.0.16 (2023-05-31)
 
 - Minor tweaks to rules documentation.